+++ packages/client-sdk/src/Mash.ts
@@ -1,13 +1,12 @@
 import { PartialDeep } from "type-fest";
 
 import * as MashAPI from "./api/routes.js";
-import { WalletButtonPosition } from "./api/routes.js";
 import parseConfig, { PartialConfig, Config } from "./config.js";
 import IFrame from "./iframe/IFrame.js";
 import {
-  FloatLocation,
   getWalletPosition,
   WalletPosition,
+  formatPosition,
 } from "./iframe/position.js";
 import MashRPCAPI, { AutopayAuthorization } from "./rpc/RPCApi.js";
 import preconnect from "./widgets/preconnect.js";
@@ -24,7 +23,7 @@ class Mash {
   private iframe: IFrame;
   private initialized = false;
   private config: Config;
-  private useInitPlacement = false;
+  private positionPromise!: Promise<MashAPI.WalletButtonPosition>;
 
   constructor(config: string | PartialConfig) {
     /**
@@ -48,14 +47,15 @@ class Mash {
       injectWidgets(this.config.widgets.baseURL);
     }
 
-    MashAPI.getEarner(this.config.api, this.config.earnerID)
+    this.positionPromise = MashAPI.getEarner(
+      this.config.api,
+      this.config.earnerID,
+    )
       .then(result => {
         if (this.config.widgets.injectTheme) {
           injectTheme(this.config.widgets.baseURL, result.customization.theme);
         }
-        if (!this.useInitPlacement) {
-          this.mountIframe(result.customization.walletButtonPosition);
-        }
+        return result.customization.walletButtonPosition;
       })
       .catch(() => {
         // If API error, inject default theme
@@ -65,9 +65,7 @@ class Mash {
             fontFamily: "inherit",
           });
         }
-        if (!this.useInitPlacement) {
-          this.mountIframe(getWalletPosition());
-        }
+        return getWalletPosition();
       });
   }
 
@@ -102,66 +100,17 @@ class Mash {
     }
 
     if (settings?.position) {
-      this.useInitPlacement = true;
-      const formattedPosition = this.formatPosition(settings?.position);
-
+      const formattedPosition = formatPosition(settings?.position);
       const position = getWalletPosition(
         formattedPosition.desktop,
         formattedPosition.mobile,
       );
-
-      return this.mountIframe(position);
-    }
-  }
-
-  formatPosition(position?: PartialDeep<WalletPosition>) {
-    const formattedPosition: WalletButtonPosition = getWalletPosition();
-    if (position?.desktop?.floatLocation === FloatLocation.BottomLeft) {
-      formattedPosition.desktop.floatSide = MashAPI.WalletButtonFloatSide.Left;
-    } else if (position?.desktop?.floatLocation === FloatLocation.BottomRight) {
-      formattedPosition.desktop.floatSide = MashAPI.WalletButtonFloatSide.Right;
+      return this.mount(position);
     }
 
-    if (position?.desktop?.shiftLeft || position?.desktop?.shiftRight) {
-      formattedPosition.desktop.floatPlacement =
-        MashAPI.WalletButtonFloatPlacement.Custom;
-      formattedPosition.desktop.customShiftConfiguration.horizontal =
-        position.desktop.shiftLeft || 0;
-      formattedPosition.desktop.customShiftConfiguration.horizontal =
-        position.desktop.shiftRight || 0;
-    }
-
-    if (position?.desktop?.shiftUp) {
-      formattedPosition.desktop.floatPlacement =
-        MashAPI.WalletButtonFloatPlacement.Custom;
-      formattedPosition.desktop.customShiftConfiguration.vertical =
-        position.desktop.shiftUp || 0;
-    }
-
-    if (position?.mobile?.floatLocation === FloatLocation.BottomLeft) {
-      formattedPosition.mobile.floatSide = MashAPI.WalletButtonFloatSide.Left;
-    } else if (position?.mobile?.floatLocation === FloatLocation.BottomRight) {
-      formattedPosition.mobile.floatSide = MashAPI.WalletButtonFloatSide.Right;
-    }
-
-    return formattedPosition;
-  }
-
-  mountIframe(position: WalletButtonPosition) {
-    return new Promise((resolve, reject) => {
-      const onIframeLoaded = (iframe: HTMLIFrameElement) => {
-        this.api = new MashRPCAPI(this.iframe.src.origin, iframe.contentWindow);
-
-        this.api
-          .init(this.config.earnerID, position)
-          .then(() => {
-            this.initialized = true;
-            resolve(null);
-          })
-          .catch(err => reject(err));
-      };
-
-      this.iframe.mount(onIframeLoaded, position);
+    return this.positionPromise.then(position => {
+      position.desktop.floatSide = MashAPI.WalletButtonFloatSide.Left;
+      this.mount(position);
     });
   }
 
@@ -223,6 +172,24 @@ class Mash {
    */
   isReady() {
     return this.initialized;
+  }
+
+  private mount(position: MashAPI.WalletButtonPosition) {
+    return new Promise((resolve, reject) => {
+      const onIframeLoaded = (iframe: HTMLIFrameElement) => {
+        this.api = new MashRPCAPI(this.iframe.src.origin, iframe.contentWindow);
+
+        this.api
+          .init(this.config.earnerID, position)
+          .then(() => {
+            this.initialized = true;
+            resolve(null);
+          })
+          .catch(err => reject(err));
+      };
+
+      this.iframe.mount(onIframeLoaded, position);
+    });
   }
 }
